InsuranceExtractionSystem/requirements.txt
# ── Core ──────────────────────────────────────────────────
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# ── Database ──────────────────────────────────────────────
sqlalchemy>=2.0.0
alembic>=1.12.0

# ── Data Processing ──────────────────────────────────────
pandas>=2.0.0
openpyxl>=3.1.0
tabulate>=0.9.0

# ── PDF Processing ───────────────────────────────────────
pypdf>=3.0.0
pdfplumber>=0.10.0

# ── AI Providers ─────────────────────────────────────────
google-generativeai>=0.8.0
openai>=1.60.0
anthropic>=0.40.0

# ── Crawler (M1) ─────────────────────────────────────────
playwright>=1.40.0
beautifulsoup4>=4.12.0

# ── SFTP (GW1) ───────────────────────────────────────────
paramiko>=3.3.0

# ── Utilities ────────────────────────────────────────────
pydantic>=2.0.0
